Remove test models and merge into existing test file
